MDX: Fixed Bug w/ Get Control Status Buffer (P4IO Timing Fix)
The `ac_io_mdxf_get_control_status_buffer` function should not have been writing to the `head` pointer. This was allowing for an edge case situation where the `head` pointer was advanced in another thread then rewound back to the previous `head` in this thread. The real implementation only reads the pointer and returns it.